% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_chs2016}
\alias{df_chs2016}
\title{Charleston, SC precinct voting data}
\format{A dat frame with columns:
\describe{
 \item{DEM}{Percent of voters who voted for the Democratic candidate}
 \item{REP}{Percent of voters who voted for the Republican candidate}
}}
\source{
scvotes.org
}
\usage{
df_chs2016
}
\description{
Charleston, SC precinct voting data
}
\keyword{datasets}

