% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf.df}
\alias{sf.df}
\title{Spatial Polygon with South Carolina Voting Districts - Data Frame}
\format{An object of class \code{data.frame} with 1250945 rows and 21 columns.}
\usage{
sf.df
}
\description{
Spatial Polygon with South Carolina Voting Districts - Data Frame
}
\keyword{datasets}

