% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_scarolina}
\alias{sp_scarolina}
\title{Spatial Polygon with South Carolina Voting Districts}
\format{An object of class \code{SpatialPolygonsDataFrame} with 2122 rows and 15 columns.}
\usage{
sp_scarolina
}
\description{
Spatial Polygon with South Carolina Voting Districts
}
\keyword{datasets}

