% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_district_choropleth.R
\docType{data}
\name{VotingDistrictChoropleth}
\alias{VotingDistrictChoropleth}
\title{An R6 object for creating Voting District -level choropleths.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VotingDistrictChoropleth
}
\description{
An R6 object for creating Voting District -level choropleths.
}
\keyword{datasets}

